feat(stg): integrate nonreturning contracts and external graph tracing
Schema 7 / execution ABI 5 carries NoSuccess without a fabricated result vector. Preserve saturation and raised-operand roots; retry raised CAFs within one invocation. Authenticate external edges through the machine ledger with pre-reserved per-copy deduplication and selected-promotion metadata.

Verified: Haskell recovered-body test; 103 prepared-program tests; 8 heap external tests, 4 promotion tests and production host-GC growth regression; workspace tests compile-only. Corpus rerun pending. Array primitive emission and Young/Retained reclamation are not complete. See plans/stg-wave5.md for evidence and trial corrections.

+++ tidepool-codegen/src/machine_state.rs
@@ -1208,7 +1208,9 @@ impl MachineState {
         let storage = self.external_storage.borrow();
         let record = storage
             .get(&published)
-            .ok_or(ExternalStorageValidationError::Untracked(published))?;
+            .ok_or(ExternalStorageValidationError::Untracked(
+                published as usize,
+            ))?;
         if record.kind != expected {
             return Err(ExternalStorageValidationError::KindMismatch {
                 expected,
@@ -1270,7 +1272,9 @@ impl MachineState {
         let storage = self.external_storage.borrow();
         for &published in marked {
             if !storage.contains_key(&published) {
-                return Err(ExternalStorageValidationError::Untracked(published));
+                return Err(ExternalStorageValidationError::Untracked(
+                    published as usize,
+                ));
             }
         }
         for (&published, record) in storage.iter() {
